upd gitpod, docker, app cfg

+++ bin/start.sh
@@ -1,9 +1,9 @@
 
 PROJECT_DIR=/workspace/my-application/
 
-cp ${PROJECT_DIR}.ops/config/env.txt ${PROJECT_DIR}.env
-
-composer install --ignore-platform-reqs
+if [[! -f ${PROJECT_DIR}.env ]]; then
+  cp ${PROJECT_DIR}.ops/config/env.txt ${PROJECT_DIR}.env
+fi
 
 docker login -u "$DOCKER_USERNAME" -p "$DOCKER_TOKEN"
 COMPOSE_FILE="${PROJECT_DIR}.gitpod.compose.yml"
